+++ CHANGES.md
@@ -4,6 +4,8 @@
 
 * Activity and Application were removed. Use View.addToDocument() instead.
 * View.addToDocument() was simplified.
+* View.outerWidth and outerHeight were renamed to realWidth and realHeight
+* DOMQuery.outerWidth, outerHeight and outerSize were renamed to width, height and size
 * Cordova is moved to another repository, rikulo_gap.
 
 **September 25, 2012**